% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitFunctions_gaussian_nlm.R
\name{fit_gaussian_nlm}
\alias{fit_gaussian_nlm}
\title{Fit a Gaussian Peak Using Non-Linear Least Squares}
\usage{
fit_gaussian_nlm(x, y, init_mu, init_sigma, init_k)
}
\arguments{
\item{x}{values of the x-axis}

\item{y}{values of the y-axis}

\item{init_mu}{initial guess of mu (the peak center)}

\item{init_sigma}{initial guess of sigma (the standard deviation)}

\item{init_k}{initial guess of k (the peak height)}
}
\value{
a list containing a vector `fitPeak` containing the y-values of the
  fitted peak and a list `nlsResult` containing the data returned from nlm()
}
\description{
Fit a Gaussian Peak Using Non-Linear Least Squares
}
\details{
Implementation uses the nls() function to fit a gaussian peak to
  experimental data using the nl2sol algorithm from the Port library - see
  ?nls() and https://people.sc.fsu.edu/~jburkardt/f_src/minpack/minpack.html
}
\examples{
 xVec <- seq(from = 1, to = 100, by = 0.1)
 gauss <- func_gaussian(x = xVec, mu = 10, sigma = 1, probDensity = FALSE, k = 10)
 gauss <- gauss + rnorm(n = length(gauss), mean = 0, sd = 1)
 fit <- fit_gaussian_nlm(x = xVec, y = gauss, init_mu = 9, init_sigma = 0.5, init_k = 3)
 plot(x = xVec, y = gauss); points(x = xVec, y = fit$fitPeak, col = "red")

}
