% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitFunctions_nGaussian_ga.R
\name{fit_nGaussian_ga}
\alias{fit_nGaussian_ga}
\title{Fit N-Gaussian Peaks Using genetic algorithm}
\usage{
fit_nGaussian_ga(x, y, init_mus, init_ks, constant_sigma, maxit = 1000)
}
\arguments{
\item{x}{values of the x-axis}

\item{y}{values of the y-axis}

\item{init_mus}{initial guess of mu (the peak center)}

\item{init_ks}{initial guess of k (the peak height)}

\item{constant_sigma}{a constand value for sigma to be applied to each of the N-peaks}

\item{maxit}{maximum number of iterations (default = 1000)}
}
\value{
a list containing a vector of y-values of the fitted peak called
  `fitPeak` and a list named `optimResult` containing the results from
  optim()
}
\description{
Fit N-Gaussian Peaks Using genetic algorithm
}
\examples{
xVec <- seq(from = 1, to = 100, by = 0.1)

}
