% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitFunctions_nGaussian_optim.R
\name{fit_nGaussian_optim}
\alias{fit_nGaussian_optim}
\title{Fit N-Gaussian Peaks Using quasi-Newtonian optomization method with boxed constraints}
\usage{
fit_nGaussian_optim(x, y, init_mus, init_ks, constant_sigma, maxit = 1000)
}
\arguments{
\item{x}{values of the x-axis}

\item{y}{values of the y-axis}

\item{init_mus}{initial guess of mu (the peak center)}

\item{init_ks}{initial guess of k (the peak height)}

\item{constant_sigma}{a constand value for sigma to be applied to each of the N-peaks}

\item{maxit}{maximum number of iterations (default = 1000)}
}
\description{
Fit N-Gaussian Peaks Using quasi-Newtonian optomization method with boxed constraints
}
