% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_func_expGaussian.R
\name{fun_expGaussian}
\alias{fun_expGaussian}
\title{Calculate an Exponentially Modified Gaussian Probability Distribution or Peak}
\usage{
fun_expGaussian(x, mu, sigma, lambda, probDist = TRUE, k)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{mu}{the mean of the gaussian component}

\item{sigma}{the standard deviation of the gaussian component}

\item{probDist}{Should the function produce a probability distribution
`TRUE` or an exponentially modified gaussian peak `FALSE` with amplitude k? default is `TRUE`.}

\item{k}{Amplitude of the gaussian peak. Only used when `probDist == TRUE`}
}
\value{
a vector of y-coordinates the same length as x
}
\description{
Calculate an Exponentially Modified Gaussian Probability Distribution or Peak
}
