% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_func_arbGaussian.R
\name{func_arbGaussian}
\alias{func_arbGaussian}
\title{Calculate Normal Distribution or Gaussian Peak}
\usage{
func_arbGaussian(x, mu, sigma, probDensity, k, weights)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{probDensity}{Should the function produce a probability density function
`TRUE` or a gaussian peak `FALSE` with amplitude k? default is `TRUE`.}

\item{k}{Amplitude of the peak. Only used when `probDensity == FALSE`}

\item{weights}{weighting vector to augment the gaussian shape profile}
}
\value{
a vector of y-coordinates the same length as x
}
\description{
Calculates either a normal distribution similar to [dnorm()] with an
integrated area of 1 or a gaussian peak with amplitude k
}
\examples{
xVec <- seq(from = 1, to = 100, by = 0.1)
weights <- rep(x = 1, times = length(xVec))
weights[490:510] <- 0.1
pdist <- func_arbGaussian(x = xVec, mu = 50, sigma = 1, probDensity = TRUE, weights = weights)
plot(xVec, pdist)
}
