% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_func_biGaussian.R
\name{func_biGaussian}
\alias{func_biGaussian}
\title{Calculate a Bi-Gaussian Probability Density or Peak}
\usage{
func_biGaussian(x, mu, sigma1, sigma2, probDensity = TRUE, k)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{mu}{the center of the peak}

\item{sigma1}{the standard deviation of the left side gaussian component}

\item{sigma2}{the standard deviation of the right side gaussian component}

\item{probDensity}{Should the function return a probability density `TRUE` or
an exponentially modified gaussian peak `FALSE` with amplitude k? default
is `TRUE`.}

\item{k}{Amplitude of the peak. Only used when `probDensity == FALSE`}
}
\value{
a vector of y-coordinates the same length as x with names "1" and "2"
  corrisponding to the two gaussian components used to generate the curve
}
\description{
Calculate a Bi-Gaussian Probability Density or Peak
}
\examples{
#Probability density of exponentially modified gaussian
xVec <- seq(from = 1, to = 100, by = 0.1)
pdist <- func_biGaussian(x = xVec, mu = 10, sigma1 = 1, sigma2 = 2, probDensity = TRUE)
plot(x = xVec, y = pdist, col = names(yVec))

#Exponentially modified gaussian peak
gpeak <- func_biGaussian(x = xVec, mu = 10, sigma1 = 1, sigma2 = 2, probDensity = FALSE, k = 10)
plot(x = xVec, y = gpeak, col = names(yVec))

}
