% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_func_expGaussian.R
\name{func_expGaussian}
\alias{func_expGaussian}
\title{Calculate an Exponentially Modified Gaussian Probability Density or Peak}
\usage{
func_expGaussian(x, mu, sigma, lambda, probDensity = TRUE, k)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{mu}{the mean of the gaussian component}

\item{sigma}{the standard deviation of the gaussian component}

\item{lambda}{rate parameter of the exponential component}

\item{probDensity}{Should the function return a probability density
`TRUE` or an exponentially modified gaussian peak `FALSE` with amplitude k? default is `TRUE`.}

\item{k}{Amplitude of the gaussian peak. Only used when `probDensity == TRUE`}
}
\value{
a vector of y-coordinates the same length as x
}
\description{
Calculate an Exponentially Modified Gaussian Probability Density or Peak
}
\examples{
#Probability density of exponentially modified gaussian
xVec <- seq(from = 1, to = 100, by = 0.1)
pdist <- func_expGaussian(x = xVec, mu = 10, sigma = 1, lambda = 0.2, probDensity = TRUE)
plot(x = xVec, y = pdist)

#Exponentially modified gaussian peak
gpeak <- func_expGaussian(x = xVec, mu = 10, sigma = 1, lambda = 0.2, probDensity = FALSE, k = 10)
plot(x = xVec, y = gpeak)

}
