% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_func_logNormal.R
\name{func_logNormal}
\alias{func_logNormal}
\title{Calculate Log-Normal Distribution or Log-Gaussian Peak}
\usage{
func_logNormal(x, mu, sigma, probDensity = TRUE, k)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{mu}{location of apex}

\item{sigma}{the standard deviation}

\item{probDensity}{Should the function produce a probability density function
`TRUE` or a gaussian peak `FALSE` with amplitude k? default is `TRUE`.}

\item{k}{Amplitude of the peak. Only used when `probDensity == FALSE`}
}
\value{
a vector of y-coordinates the same length as x
}
\description{
Calculates either a normal distribution with an
integrated area of 1 or a gaussian peak with amplitude k
}
\examples{
#normal distribution
xVec <- seq(from = 1, to = 100, by = 0.1)
pdensity <- func_gaussian(x = xVec, mu = 10, sigma = 1, probDensity = T)
p1 <- plot(x = xVec, y = pdensity)

#gaussian peak
gpeak <- func_gaussian(x = xVec, mu = 10, sigma = 1, probDensity = F, k = 10)
p2 <- plot(x = xVec, y = gpeak)

#See also: dnorm()
}
