% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_func_lorentzian.R
\name{func_lorentzian}
\alias{func_lorentzian}
\title{Calculate Cauchy (Lorentz) probability density function or a Lorentzian Peak}
\usage{
func_lorentzian(x, x0, gamma, probDensity = TRUE, k)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{x0}{location parameter specifying the location of the peak of the distribution}

\item{gamma}{scale parameter specifying the half-width at half-maximum (HWHM)}

\item{probDensity}{Should the function produce a probability density function
`TRUE` or a gaussian peak `FALSE` with amplitude k? default is `TRUE`.}

\item{k}{amplitude of the peak at location x0}
}
\value{
a vector of y-coordinates the same length as x
}
\description{
Calculate Cauchy (Lorentz) probability density function or a Lorentzian Peak
}
\examples{
#Probability density function
xVec <- seq(from = 1, to = 100, by = 0.1)
pdist <- func_lorentzian(x = xVec, x0 = 50, gamma = 1, probDensity = TRUE)
plot(x = xVec, y = pdist)
#Make a peak
lpeak <- func_lorentzian(x = xVec, x0 = 50, gamma = 1, probDensity = FALSE, k = 10)
plot(x = xVec, y = lpeak)
}
