% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_func_pseudoVoigt.R
\name{func_pseudoVoigt}
\alias{func_pseudoVoigt}
\title{Calculate Pseudo-Voigt Using Numerical Approximation}
\usage{
func_pseudoVoigt(x, sigma, gamma, mu, probDensity, k)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{sigma}{the standard deviation of the gaussian}

\item{gamma}{the lorentzian scale parameter specifying the half-width at half-maximum (HWHM)}

\item{mu}{the mean}

\item{probDensity}{Should the function produce a probability density function
`TRUE` a peak `FALSE` with amplitude k? default is `TRUE`.}

\item{k}{Amplitude of the peak. Only used when `probDensity == FALSE`}
}
\value{
a vector of y-coordinates the same length as x
}
\description{
Calculates the linear combination of gaussian and lorentzian peaks with normalization factor
}
\examples{
xVec <- seq(from = 1, to = 100, by = 0.1)
pdensity <- func_pseudoVoigt(x = xVec, mu = 10, sigma = 1, gamma = 1.5, probDensity = TRUE)
p1 <- plot(x = xVec, y = pdensity)

#Pseudo-Voigt Peak
pdensity <- func_pseudoVoigt(x = xVec, mu = 10, sigma = 1, gamma = 1.5, probDensity = FALSE, k = 10)
p2 <- plot(x = xVec, y = pdensity)
}
