% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMax.R
\name{localMax}
\alias{localMax}
\title{Calculate local maximum within a local neighborhood}
\usage{
localMax(vec, neighborhood)
}
\arguments{
\item{vec}{a vector to be evaluated for the presence of local maxima}

\item{neighborhood}{number of points to the left and right of a point for
calculating local maxima (see details).}
}
\value{
a logical vector masking the positions of local maxima
}
\description{
Calculate local maximum within a local neighborhood
}
\details{
`neighborhood = 1` will detect local maxima at position `i` defind
  as a point that is greater than both its neighbors to the left `(i-1)` and
  right `(i+1)`. Setting `neighborhood > 1` results in detecting local maxima
  at a position `i` defined as a point that is greater than all neighboring
  points to the left `c(i-1, i-..., i-n)` and right `c(i+1, i+..., i+n)`.
}
\examples{
dat <- rnorm(n = 100, mean = 0, sd = 1)
i <- seq_along(dat)
locMax <- localMax(vec = dat, neighborhood = 1)

i_locMax <- i[locMax]
locMax <-dat[locMax]

plot(x = i, y = dat); points(x = i_locMax, y = locMax, col = "red")

}
