% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMin.R
\name{localMin}
\alias{localMin}
\title{Calculate local minimum within a local neighborhood}
\usage{
localMin(vec, neighborhood)
}
\arguments{
\item{vec}{a vector to be evaluated for the presence of local minimum}

\item{neighborhood}{number of points to the left and right of a point for
calculating local minima (see details).}
}
\value{
a logical vector masking the positions of local minima
}
\description{
Calculate local minimum within a local neighborhood
}
\details{
`neighborhood = 1` will detect local minima at position `i` defind
  as a point that is less than both its neighbors to the left `(i-1)` and
  right `(i+1)`. Setting `neighborhood > 1` results in detecting local minima
  at a position `i` defined as a point that is less than all neighboring
  points to the left `c(i-1, i-..., i-n)` and right `c(i+1, i+..., i+n)`.
}
\examples{
dat <- rnorm(n = 100, mean = 0, sd = 1)
i <- seq_along(dat)
locMin <- localMin(vec = dat, neighborhood = 1)

i_locMin <- i[locMin]
locMin <-dat[locMin]

plot(x = i, y = dat); points(x = i_locMin, y = locMin, col = "red")

}
