% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_multi_lorentzian.R
\name{multi_lorentzian}
\alias{multi_lorentzian}
\title{Calculate Cauchy (Lorentz) probability density function(s) or Lorentzian Peak(s)}
\usage{
multi_lorentzian(x, x0s, gammas, probDensity, returnComponentPks = FALSE, ks)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{x0s}{location parameter specifying the locations of the peaks of the distribution}

\item{gammas}{scale parameter specifying the half-widths at half-maximum (HWHM)}

\item{probDensity}{Should the function produce a probability density function
`TRUE` or a gaussian peak `FALSE` with amplitude k? default is `TRUE`.}

\item{returnComponentPks}{Should the function return a single vector
containing the sum of each individual peak `FALSE` or a data.frame
containing the input vector `x`, each component peak `peak_n`, and the
summed result of the component peaks `peak_sum`}

\item{ks}{amplitude of the peak at location x0}
}
\value{
returns either a single vector of y-coordinates the same length as x or a data.frame
}
\description{
Calculate Cauchy (Lorentz) probability density function(s) or Lorentzian Peak(s)
}
\examples{
xVec <- seq(from = 1, to = 100, by = 0.1)
pdist <- multi_lorentzian(x = xVec,
                          x0s = c(5, 10, 15),
                          gammas = c(1, 2, 4),
                          probDensity = TRUE,
                          returnComponentPks = FALSE)

plot(x = xVec, y = pdist)

lpeak_components <- multi_lorentzian(x = xVec,
                                     x0s = c(5, 10, 15),
                                     gammas = c(1, 2, 4),
                                     probDensity = FALSE,
                                     returnComponentPks = TRUE,
                                     ks = c(1, 5, 10))

plot(x = lpeak_components$x, y = lpeak_components$peak_sum)
points(x = lpeak_components$x, y = lpeak_components$peak_1, col = "red")
points(x = lpeak_components$x, y = lpeak_components$peak_2, col = "blue")
points(x = lpeak_components$x, y = lpeak_components$peak_3, col = "green")

}
