% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_multiPeak_sum.R
\name{multiPeak_sum}
\alias{multiPeak_sum}
\title{Formatting/housekeeping function for PeakFunctions output}
\usage{
multiPeak_sum(x, peaks, probDensity, returnComponentPks)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{peaks}{the peak or peaks to be summed}

\item{probDensity}{Should the function produce a probability density function
`TRUE` or a gaussian peak `FALSE` with amplitude k? default is `TRUE`.}

\item{returnComponentPks}{Should the function return a single vector
containing the sum of each individual peak `FALSE` or a data.frame
containing the input vector `x`, each component peak `peak_n`, and the
summed result of the component peaks `peak_sum`}
}
\value{
either a vector or data.frame
}
\description{
Formatting/housekeeping function for PeakFunctions output
}
