% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMax.R
\name{rightShift_greater}
\alias{rightShift_greater}
\title{Check X-right side neighbors are greater}
\usage{
rightShift_greater(X, vec, vecLength)
}
\arguments{
\item{X}{a integer vector of sequence c(1:neighborhood)}

\item{vec}{the vector of values to be evaluated}

\item{vecLength}{the length of vec}
}
\value{
logical vector padded with `FALSE` values on the end
}
\description{
Check X-right side neighbors are greater
}
