% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitFunctions_gaussian_optim.R
\name{fit_gaussian_optim}
\alias{fit_gaussian_optim}
\title{Fit a Gaussian Peak Using quasi-Newtonian optomization method with boxed
constraints}
\usage{
fit_gaussian_optim(x, y, init_mu, init_sigma, init_k, maxit = 1000)
}
\arguments{
\item{x}{values of the x-axis}

\item{y}{values of the y-axis}

\item{init_mu}{initial guess of mu (the peak center)}

\item{init_sigma}{initial guess of sigma (the standard deviation)}

\item{init_k}{initial guess of k (the peak height)}

\item{maxit}{maximum number of iterations (default = 1000)}
}
\value{
a list containing a vector of y-values of the fitted peak called
  `fitPeak` and a list named `optimResult` containing the results from
  optim()
}
\description{
Fit a Gaussian Peak Using quasi-Newtonian optomization method with boxed
constraints
}
