% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FitFunctions_nGaussian_ga.R
\name{fit_nGaussian_ga}
\alias{fit_nGaussian_ga}
\title{Fit N-Gaussian Peaks Using genetic algorithm}
\usage{
fit_nGaussian_ga(
  x,
  y,
  init_mus,
  init_ks,
  init_sigmas,
  lower_mus,
  lower_ks,
  lower_sigmas,
  upper_mus,
  upper_ks,
  upper_sigmas,
  maxit = 1000
)
}
\arguments{
\item{x}{values of the x-axis to be fit}

\item{y}{values of the y-axis to be fit}

\item{init_mus}{initial guess of mu (the peak center) - either a single value to be applied globally or a value specifying each peak}

\item{init_ks}{initial guess of k (the peak height)}

\item{maxit}{}
}
\value{

}
\description{
Fit N-Gaussian Peaks Using genetic algorithm
}
