% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_func_arbGaussian.R
\name{func_arbGaussian}
\alias{func_arbGaussian}
\title{Calculate Normal Distribution or Gaussian Peak}
\usage{
func_arbGaussian(x, mu, sigma, probDensity, k, weights)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{probDensity}{Should the function produce a probability density function
`TRUE` or a gaussian peak `FALSE` with amplitude k? default is `TRUE`.}

\item{k}{Amplitude of the peak. Only used when `probDensity == FALSE`}

\item{weights}{weighting vector to augment the gaussian shape profile}
}
\value{
a vector of y-coordinates the same length as x
}
\description{
Calculates either a normal distribution similar to [dnorm()] with an
integrated area of 1 or a gaussian peak with amplitude k
}
\examples{
#normal distribution
xVec <- seq(from = 1, to = 100, by = 0.1)
pdensity <- func_gaussian(x = xVec, mu = 10, sigma = 1, probDensity = T)
p1 <- plot(x = xVec, y = pdensity)

#gaussian peak
gpeak <- func_gaussian(x = xVec, mu = 10, sigma = 1, probDensity = F, k = 10)
p2 <- plot(x = xVec, y = gpeak)

#See also: dnorm()
}
