% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PeakFunctions_func_gaussian.R
\name{func_gaussian}
\alias{func_gaussian}
\title{Calculate Normal Distribution or Gaussian Peak}
\usage{
func_gaussian(x, mu, sigma, probDist = TRUE, k)
}
\arguments{
\item{x}{a vector of x-coordinates from which the corrisponding y-coordinates
are calculated}

\item{mu}{the mean}

\item{sigma}{the standard deviation}

\item{probDist}{Should the function produce a probability distribution
`TRUE` or a gaussian peak `FALSE` with amplitude k? default is `TRUE`.}

\item{k}{Amplitude of the gaussian peak. Only used when `probDist == TRUE`}
}
\value{
a vector of y-coordinates the same length as x
}
\description{
Calculates either a normal distribution similar to [dnorm()] with an
integrated area of 1 or a gaussian peak with amplitude k
}
\examples{
#normal distribution
xVec <- seq(from = 1, to = 100, by = 0.1)
pdensity <- func_gaussian(x = xVec, mu = 10, sigma = 1, probDist = T)
p1 <- plot(x = xVec, y = pdensity)

#gaussian peak
gpeak <- func_gaussian(x = xVec, mu = 10, sigma = 1, probDist = F, k = 10)
p2 <- plot(x = xVec, y = gpeak)

#See also: dnorm()
}
